--- .github/workflows/update_llm_perf_intel_pytorch.yml
@@ -1,56 +1,0 @@
-name: Update LLM Perf Benchmarks - Intel PyTorch
-
-on:
-  workflow_dispatch:
-  schedule:
-    - cron: "0 0 * * *"
-
-concurrency:
-  cancel-in-progress: true
-  group: ${{ github.workflow }}-${{ github.ref }}
-
-env:
-  IMAGE: ghcr.io/huggingface/optimum-benchmark:latest-cpu
-
-jobs:
-  run_benchmarks:
-    strategy:
-      fail-fast: false
-      matrix:
-        backend: [pytorch, onnxruntime, openvino]
-        subset: [unquantized]
-        machine: [
-          {name: c7i, runs-on: {group: 'aws-c7i-8xlarge-plus'}}, 
-        ]
-
-    runs-on: ${{ matrix.machine.runs-on }}
-
-    steps:
-      - name: Checkout
-        uses: actions/checkout@v4
-
-      - name: Run benchmarks
-        uses: addnab/docker-run-action@v3
-        env:
-          BACKEND: ${{ matrix.backend }}
-          SUBSET: ${{ matrix.subset }}
-          MACHINE: ${{ matrix.machine.name }}
-          HF_TOKEN: ${{ secrets.HF_TOKEN }}
-        with:
-          image: ${{ env.IMAGE }}
-          options: |
-            --rm
-            --shm-size 64G
-            --env BACKEND
-            --env SUBSET
-            --env MACHINE
-            --env HF_TOKEN
-            --env MKL_THREADING_LAYER=GNU
-            --env HF_HUB_ENABLE_HF_TRANSFER=1
-            --volume ${{ github.workspace }}:/workspace
-            --workdir /workspace
-          run: |
-            pip install packaging && pip install einops scipy optimum codecarbon
-            pip install -U transformers huggingface_hub[hf_transfer]
-            pip install -e .[onnxruntime,openvino]
-            python llm_perf/hardware/intel/update_llm_perf_intel.py
