+++ llm_perf/hardware.yml
@@ -1,6 +1,7 @@
 - machine: 1xA10
   description: A10-24GB-150W 🖥️
-  hardware_type: cuda
+  hardware_provider: nvidia
+  hardware_type: gpu
   subsets:
     - unquantized
     - awq
@@ -11,7 +12,8 @@
 
 - machine: 1xA100
   description: A100-80GB-275W 🖥️
-  hardware_type: cuda
+  hardware_provider: nvidia
+  hardware_type: gpu
   subsets:
     - unquantized
     - awq
@@ -22,7 +24,8 @@
 
 - machine: 1xT4
   description: T4-16GB-70W 🖥️
-  hardware_type: cuda
+  hardware_provider: nvidia
+  hardware_type: gpu
   subsets:
     - unquantized
     - awq
@@ -33,7 +36,8 @@
 
 - machine: c7i
   description: 4th-Gen-Intel-Xeon-385W 🖥️
-  hardware_type: intel
+  hardware_provider: intel
+  hardware_type: cpu
   subsets:
     - unquantized
   backends:
